+++ project/src/hooks/use-login/useLogin.js
@@ -1,0 +1,17 @@
+import {useRef} from 'react';
+
+export const useLogin = (onSubmit) => {
+  const loginRef = useRef();
+  const passwordRef = useRef();
+
+  const handleSubmit = (e) => {
+    e.preventDefault();
+
+    onSubmit({
+      login: loginRef.current.value,
+      password: passwordRef.current.value,
+    });
+  };
+
+  return [loginRef, passwordRef, handleSubmit];
+};
